Move SheetExtentScope to separate file
