+++ package/lib/src/draggable/draggable_sheet_extent_scope.dart
@@ -1,6 +1,7 @@
 import 'package:meta/meta.dart';
 
 import '../foundation/sheet_extent.dart';
+import '../foundation/sheet_extent_scope.dart';
 import 'draggable_sheet_extent.dart';
 
 @internal
