+++ package/lib/src/draggable/sheet_draggable.dart
@@ -3,6 +3,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/widgets.dart';
 
 import '../foundation/sheet_extent.dart';
+import '../foundation/sheet_extent_scope.dart';
 import '../scrollable/scrollable_sheet.dart';
 import 'draggable_sheet.dart';
