+++ package/lib/src/foundation/sheet_content_scaffold.dart
@@ -5,6 +5,7 @@ import 'package:flutter/rendering.dart';
 
 import '../draggable/sheet_draggable.dart';
 import 'sheet_extent.dart';
+import 'sheet_extent_scope.dart';
 import 'sheet_viewport.dart';
 
 class SheetContentScaffold extends StatelessWidget {
