+++ package/lib/src/foundation/sheet_viewport.dart
@@ -5,6 +5,7 @@ import 'package:flutter/widgets.dart';
 import 'package:meta/meta.dart';
 
 import 'sheet_extent.dart';
+import 'sheet_extent_scope.dart';
 
 @internal
 class SheetViewport extends SingleChildRenderObjectWidget {
