+++ package/lib/src/navigation/navigation_sheet.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 
 import '../foundation/keyboard_dismissible.dart';
 import '../foundation/sheet_controller.dart';
-import '../foundation/sheet_extent.dart';
+import '../foundation/sheet_extent_scope.dart';
 import '../foundation/sheet_gesture_tamperer.dart';
 import '../foundation/sheet_theme.dart';
 import '../foundation/sheet_viewport.dart';
