+++ package/lib/src/navigation/navigation_sheet_extent_scope.dart
@@ -1,6 +1,7 @@
 import 'package:meta/meta.dart';
 
 import '../foundation/sheet_extent.dart';
+import '../foundation/sheet_extent_scope.dart';
 import '../foundation/sheet_physics.dart';
 import 'navigation_sheet_extent.dart';
