+++ package/lib/src/scrollable/scrollable_sheet_extent_scope.dart
@@ -1,6 +1,7 @@
 import 'package:meta/meta.dart';
 
 import '../foundation/sheet_extent.dart';
+import '../foundation/sheet_extent_scope.dart';
 import 'scrollable_sheet_extent.dart';
 
 @internal
