+++ package/lib/src/scrollable/sheet_scrollable.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
-import '../foundation/sheet_extent.dart';
+import '../foundation/sheet_extent_scope.dart';
 import 'scrollable_sheet_extent.dart';
 import 'sheet_content_scroll_position.dart';
